// src/api/model/organization.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MaCarte API
 * Documentation OpenAPI de l'API MaCarte
 * OpenAPI spec version: 1.0.0
 */
import type { OrganizationMembersItem } from './organizationMembersItem';

export interface Organization {
  /** Identifiant de l'organisation */
  public_id?: string;
  /** Nom de l'organisation */
  name?: string;
  /**
     * Présentation / Description de l'organisation
     * @nullable
     */
  presentation?: string | null;
  /**
     * URL de l'image / logo de l'organisation
     * @nullable
     */
  profile_picture?: string | null;
  /**
     * URL de l'image / logo de couverture de l'organisation
     * @nullable
     */
  cover_picture?: string | null;
  /** Date de création de l'organisation */
  created_at?: string;
  /** Date de modification de l'organisation */
  updated_at?: string;
  /**
     * Premium l'organisation parmi [null, 'edugeo']
     * @nullable
     */
  premium?: string | null;
  /** Nombre de membres dans l'organisation */
  nb_members?: number;
  /** Nombre de cartes de l'organisation partagées dans l'atlas */
  nb_shared_maps?: number;
  /** Nombre de vues sur l'ensemble des cartes de l'organisation */
  nb_views?: number;
  /** L'organisation est bloquée par un admin */
  locked?: boolean;
  /** Taille globale autorisée pour les images de l'organisation (en octets) */
  media_size_limit?: number;
  /** Taille globale des images de l'organisation (en octets) */
  medias_size?: number;
  /** Membres de l'organisation */
  members?: OrganizationMembersItem[];
}